pyproject.toml
[project]
name = "server-fit-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib==3.0.1",
    "asgiref>=3.11.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "boto3>=1.42.39",
    "fastapi>=0.128.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.4",
    "openai>=2.16.0",
    "pdfplumber>=0.11.9",
    "pgvector>=0.4.2",
    "psutil>=7.2.2",
    "psycopg>=3.3.2",
    "psycopg-pool>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "torch",
    "uvicorn[standard]>=0.40.0",
]
[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
explicit = true  # ← Critical: only use this index for packages that explicitly request it
